+++ src/application/hooks/api/index.ts
@@ -1,4 +1,3 @@
 export * from "./account";
 export * from "./collection";
 export * from "./search";
-export * from "./tags";
